developed creature hierarchy
One thing I noticed was that none of the classes have parameters, however Creature is abstract and as a result when demon subclasses try to input their strength and health, if Demon does not have parameters, the subclasses produce an error of arguments. small tweak but it doesn't seem to effect the system. Also I was able to combine damage and attack since they are so closely intertwined anyway.